Always use queue.h from compat/.
Was going to fix a missing include of config.h in subscription.h but that
ended up breaking stuff. The glibc version of queue.h does not have
LIST_FOREACH_SAFE (at least with glibc 2.15), which is used in
subscription.c. So instead of adding the missing include and writing a
test whether LIST_FOREACH_SAFE is available, always use queue.h from compat.